WIP: Create MessageService (saveMessage, addMessage, submitMessage, handleMessage)
