+++ src/components/context/MessageDB.tsx
@@ -15,6 +15,7 @@ import {
   MessageDB,
   UserConfig,
 } from '../../db/messages';
+import { MessageService } from '../../services/MessageService';
 import {
   buildConversationsKey,
   buildMessagesKey,
@@ -252,204 +253,7 @@ const MessageDBProvider: FC<MessageDBContextProps> = ({ children }) => {
     };
   }>({});
 
-  const saveMessage = async (
-    decryptedContent: Message,
-    messageDB: MessageDB,
-    spaceId: string,
-    channelId: string,
-    conversationType: string,
-    updatedUserProfile: { user_icon?: string; display_name?: string }
-  ) => {
-    if (decryptedContent.content.type === 'reaction') {
-      const reaction = decryptedContent.content as ReactionMessage;
-      const target = await messageDB.getMessage({
-        spaceId: spaceId,
-        channelId: channelId,
-        messageId: decryptedContent.content.messageId,
-      });
-      if (target) {
-        const existing = target.reactions?.find(
-          (r) => r.emojiId === reaction.reaction
-        );
-        const modifiedSet = [
-          ...(existing?.memberIds ?? []).filter((e) => e !== reaction.senderId),
-          reaction.senderId,
-        ];
-        await messageDB.saveMessage(
-          {
-            ...target,
-            reactions: [
-              ...(target.reactions?.filter(
-                (r) => r.emojiId !== reaction.reaction
-              ) ?? []),
-              {
-                emojiId: reaction.reaction,
-                emojiName: reaction.reaction,
-                spaceId: spaceId != channelId ? spaceId : '',
-                count: modifiedSet.length,
-                memberIds: modifiedSet,
-              },
-            ],
-          },
-          0,
-          spaceId,
-          conversationType,
-          updatedUserProfile.user_icon!,
-          updatedUserProfile.display_name!
-        );
-      } else {
-        return;
-      }
-    } else if (decryptedContent.content.type === 'remove-reaction') {
-      const reaction = decryptedContent.content as RemoveReactionMessage;
-      const target = await messageDB.getMessage({
-        spaceId: spaceId,
-        channelId: channelId,
-        messageId: decryptedContent.content.messageId,
-      });
-      if (target) {
-        const existing = target.reactions?.find(
-          (r) => r.emojiId === reaction.reaction
-        );
-        if (existing) {
-          const modifiedSet = [
-            ...(existing?.memberIds ?? []).filter(
-              (e) => e !== reaction.senderId
-            ),
-          ];
-          let reactions =
-            target.reactions?.filter((r) => r.emojiId !== reaction.reaction) ??
-            [];
-          if (modifiedSet.length != 0) {
-            reactions = [
-              ...reactions,
-              {
-                emojiId: reaction.reaction,
-                emojiName: reaction.reaction,
-                spaceId: spaceId != channelId ? spaceId : '',
-                count: modifiedSet.length,
-                memberIds: modifiedSet,
-              },
-            ];
-          }
-          await messageDB.saveMessage(
-            {
-              ...target,
-              reactions: reactions,
-            },
-            0,
-            spaceId,
-            conversationType,
-            updatedUserProfile.user_icon!,
-            updatedUserProfile.display_name!
-          );
-        }
-      } else {
-        return;
-      }
-    } else if (decryptedContent.content.type === 'remove-message') {
-      const targetMessage = await messageDB.getMessage({
-        spaceId,
-        channelId,
-        messageId: decryptedContent.content.removeMessageId,
-      });
-      if (!targetMessage) {
-        return;
-      }
-
-      if (
-        targetMessage.channelId !== decryptedContent.channelId ||
-        targetMessage.spaceId !== decryptedContent.spaceId
-      ) {
-        return;
-      }
-
-      if (
-        targetMessage.content.senderId === decryptedContent.content.senderId
-      ) {
-        await messageDB.deleteMessage(decryptedContent.content.removeMessageId);
-        return;
-      }
-
-      if (spaceId != channelId) {
-        const space = await messageDB.getSpace(spaceId);
-
-        // For read-only channels: ISOLATED permission system - only managers can delete
-        const channel = space?.groups
-          ?.find((g) => g.channels.find((c) => c.channelId === channelId))
-          ?.channels.find((c) => c.channelId === channelId);
-
-        if (channel?.isReadOnly) {
-          const isManager = !!(
-            channel.managerRoleIds &&
-            space?.roles?.some(
-              (role) =>
-                channel.managerRoleIds?.includes(role.roleId) &&
-                role.members.includes(decryptedContent.content.senderId)
-            )
-          );
-          if (isManager) {
-            await messageDB.deleteMessage(
-              decryptedContent.content.removeMessageId
-            );
-            return;
-          }
-          // For read-only channels, if not a manager, deny delete (even if user has traditional roles)
-          return;
-        }
-
-        // For regular channels: Traditional role-based permissions
-        if (
-          !space?.roles.find(
-            (r) =>
-              r.members.includes(decryptedContent.content.senderId) &&
-              r.permissions.includes('message:delete')
-          )
-        ) {
-          return;
-        }
-        await messageDB.deleteMessage(decryptedContent.content.removeMessageId);
-      }
-    } else if (decryptedContent.content.type === 'update-profile') {
-      const participant = await messageDB.getSpaceMember(
-        decryptedContent.spaceId,
-        decryptedContent.content.senderId
-      );
-      if (
-        !participant ||
-        !decryptedContent.publicKey ||
-        !decryptedContent.signature
-      ) {
-        return;
-      }
-
-      const sh = await sha256.digest(
-        Buffer.from(decryptedContent.publicKey, 'hex')
-      );
-      const inboxAddress = base58btc.baseEncode(sh.bytes);
-
-      if (
-        participant.inbox_address &&
-        participant.inbox_address != inboxAddress
-      ) {
-        return;
-      }
-
-      participant.display_name = decryptedContent.content.displayName;
-      participant.user_icon = decryptedContent.content.userIcon;
-      participant.inbox_address = inboxAddress;
-      await messageDB.saveSpaceMember(decryptedContent.spaceId, participant);
-    } else {
-      await messageDB.saveMessage(
-        { ...decryptedContent, channelId: channelId, spaceId: spaceId },
-        0,
-        spaceId,
-        conversationType,
-        updatedUserProfile.user_icon!,
-        updatedUserProfile.display_name!
-      );
-    }
-  };
+  // saveMessage will be defined after messageService instantiation
 
   const deleteEncryptionStates = React.useCallback(
     async ({ conversationId }: { conversationId: string }) => {
@@ -471,282 +275,7 @@ const MessageDBProvider: FC<MessageDBContextProps> = ({ children }) => {
     [messageDB]
   );
 
-  const addMessage = async (
-    queryClient: QueryClient,
-    spaceId: string,
-    channelId: string,
-    decryptedContent: Message
-  ) => {
-    if (decryptedContent.content.type === 'reaction') {
-      const reaction = decryptedContent.content as ReactionMessage;
-      queryClient.setQueryData(
-        buildMessagesKey({ spaceId: spaceId, channelId: channelId }),
-        (oldData: InfiniteData<any>) => {
-          if (!oldData?.pages) return oldData;
-
-          return {
-            pageParams: oldData.pageParams,
-            pages: oldData.pages.map((page, index) => {
-              return {
-                ...page,
-                messages: [
-                  ...page.messages.map((m: Message) => {
-                    if (m.messageId === reaction.messageId) {
-                      const existing = m.reactions?.find(
-                        (r) => r.emojiId === reaction.reaction
-                      );
-                      const modifiedSet = [
-                        ...(existing?.memberIds ?? []).filter(
-                          (e) => e !== reaction.senderId
-                        ),
-                        reaction.senderId,
-                      ];
-                      return {
-                        ...m,
-                        reactions: [
-                          ...(m.reactions?.filter(
-                            (r) => r.emojiId !== reaction.reaction
-                          ) ?? []),
-                          {
-                            emojiId: reaction.reaction,
-                            emojiName: reaction.reaction,
-                            spaceId: spaceId !== channelId ? spaceId : '',
-                            count: modifiedSet.length,
-                            memberIds: modifiedSet,
-                          },
-                        ],
-                      };
-                    }
-                    return m;
-                  }),
-                ],
-                // Preserve any cursors or other pagination metadata
-                nextCursor: page.nextCursor,
-                prevCursor: page.prevCursor,
-              };
-            }),
-          };
-        }
-      );
-    } else if (decryptedContent.content.type === 'remove-reaction') {
-      const reaction = decryptedContent.content as RemoveReactionMessage;
-      queryClient.setQueryData(
-        buildMessagesKey({ spaceId: spaceId, channelId: channelId }),
-        (oldData: InfiniteData<any>) => {
-          if (!oldData?.pages) return oldData;
-
-          return {
-            pageParams: oldData.pageParams,
-            pages: oldData.pages.map((page, index) => {
-              return {
-                ...page,
-                messages: [
-                  ...page.messages.map((m: Message) => {
-                    if (m.messageId === reaction.messageId) {
-                      const existing = m.reactions?.find(
-                        (r) => r.emojiId === reaction.reaction
-                      );
-                      if (existing) {
-                        const modifiedSet = [
-                          ...(existing?.memberIds ?? []).filter(
-                            (e) => e !== reaction.senderId
-                          ),
-                        ];
-                        let reactions =
-                          m.reactions?.filter(
-                            (r) => r.emojiId !== reaction.reaction
-                          ) ?? [];
-                        if (modifiedSet.length != 0) {
-                          reactions = [
-                            ...reactions,
-                            {
-                              emojiId: reaction.reaction,
-                              emojiName: reaction.reaction,
-                              spaceId: spaceId != channelId ? spaceId : '',
-                              count: modifiedSet.length,
-                              memberIds: modifiedSet,
-                            },
-                          ];
-                        }
-                        return {
-                          ...m,
-                          reactions: reactions,
-                        };
-                      }
-                      return m;
-                    }
-                    return m;
-                  }),
-                ],
-                // Preserve any cursors or other pagination metadata
-                nextCursor: page.nextCursor,
-                prevCursor: page.prevCursor,
-              };
-            }),
-          };
-        }
-      );
-    } else if (decryptedContent.content.type === 'remove-message') {
-      const targetMessage = await messageDB.getMessage({
-        spaceId,
-        channelId,
-        messageId: decryptedContent.content.removeMessageId,
-      });
-
-      // Check if this delete request should be honored
-      let shouldHonorDelete = false;
-
-      if (!targetMessage) {
-        // If target message doesn't exist, always remove from UI
-        shouldHonorDelete = true;
-      } else {
-        // If target message exists, check permissions
-
-        // 1. Users can always delete their own messages
-        if (
-          targetMessage.content.senderId === decryptedContent.content.senderId
-        ) {
-          shouldHonorDelete = true;
-        } else {
-          if (!shouldHonorDelete && spaceId != channelId) {
-            const space = await messageDB.getSpace(spaceId);
-
-            // 3. Check read-only channel manager privileges
-            const channel = space?.groups
-              ?.find((g) => g.channels.find((c) => c.channelId === channelId))
-              ?.channels.find((c) => c.channelId === channelId);
-
-            if (channel?.isReadOnly && channel.managerRoleIds) {
-              const isManager = space?.roles?.some(
-                (role) =>
-                  channel.managerRoleIds?.includes(role.roleId) &&
-                  role.members.includes(decryptedContent.content.senderId)
-              );
-              if (isManager) {
-                shouldHonorDelete = true;
-                console.log(
-                  '🔹 ADDMESSAGE: Honoring read-only manager delete in UI cache'
-                );
-              }
-            }
-
-            // 4. Check traditional role permissions
-            if (!shouldHonorDelete && !channel?.isReadOnly) {
-              const hasDeleteRole = space?.roles?.find(
-                (r) =>
-                  r.members.includes(decryptedContent.content.senderId) &&
-                  r.permissions.includes('message:delete')
-              );
-              if (hasDeleteRole) {
-                shouldHonorDelete = true;
-                console.log(
-                  '🔹 ADDMESSAGE: Honoring role-based delete in UI cache'
-                );
-              }
-            }
-          }
-        }
-      }
-
-      if (shouldHonorDelete) {
-        const targetId = decryptedContent.content.removeMessageId;
-        queryClient.setQueryData(
-          buildMessagesKey({ spaceId: spaceId, channelId: channelId }),
-          (oldData: InfiniteData<any>) => {
-            if (!oldData?.pages) return oldData;
-
-            return {
-              pageParams: oldData.pageParams,
-              pages: oldData.pages.map((page, index) => {
-                return {
-                  ...page,
-                  messages: [
-                    ...page.messages.filter(
-                      (m: Message) => m.messageId !== targetId
-                    ),
-                  ],
-                  // Preserve any cursors or other pagination metadata
-                  nextCursor: page.nextCursor,
-                  prevCursor: page.prevCursor,
-                };
-              }),
-            };
-          }
-        );
-      } else {
-        console.log('🔹 ADDMESSAGE: Ignoring unauthorized delete request');
-      }
-    } else if (decryptedContent.content.type === 'update-profile') {
-      const participant = await messageDB.getSpaceMember(
-        decryptedContent.spaceId,
-        decryptedContent.content.senderId
-      );
-      if (
-        !participant ||
-        !decryptedContent.publicKey ||
-        !decryptedContent.signature
-      ) {
-        return;
-      }
-
-      const sh = await sha256.digest(
-        Buffer.from(decryptedContent.publicKey, 'hex')
-      );
-      const inboxAddress = base58btc.baseEncode(sh.bytes);
-
-      if (
-        participant.inbox_address &&
-        participant.inbox_address != inboxAddress
-      ) {
-        return;
-      }
-
-      participant.display_name = decryptedContent.content.displayName;
-      participant.user_icon = decryptedContent.content.userIcon;
-      participant.inbox_address = inboxAddress;
-      await queryClient.setQueryData(
-        buildSpaceMembersKey({ spaceId: decryptedContent.spaceId }),
-        (oldData: secureChannel.UserProfile[]) => {
-          return [
-            ...(oldData ?? []).filter(
-              (p) => p.user_address !== participant.user_address
-            ),
-            participant,
-          ];
-        }
-      );
-    } else {
-      queryClient.setQueryData(
-        buildMessagesKey({ spaceId: spaceId, channelId: channelId }),
-        (oldData: InfiniteData<any>) => {
-          if (!oldData?.pages) return oldData;
-
-          return {
-            pageParams: oldData.pageParams,
-            pages: oldData.pages.map((page, index) => {
-              // Only add the new message to the most recent page
-              if (index === oldData.pages.length - 1) {
-                return {
-                  ...page,
-                  messages: [
-                    ...page.messages.filter(
-                      (m: Message) => m.messageId !== decryptedContent.messageId
-                    ),
-                    decryptedContent,
-                  ],
-                  // Preserve any cursors or other pagination metadata
-                  nextCursor: page.nextCursor,
-                  prevCursor: page.prevCursor,
-                };
-              }
-              // Return other pages unchanged
-              return page;
-            }),
-          };
-        }
-      );
-    }
-  };
+  // addMessage will be defined after messageService instantiation
 
   const deleteInboxMessages = async (
     inboxKeyset: secureChannel.InboxKeyset,
@@ -840,1330 +369,8 @@ const MessageDBProvider: FC<MessageDBContextProps> = ({ children }) => {
     invalidateConversation({ conversationId });
   };
 
-  const handleNewMessage = useCallback(
-    async (
-      self_address: string,
-      keyset: {
-        userKeyset: secureChannel.UserKeyset;
-        deviceKeyset: secureChannel.DeviceKeyset;
-      },
-      message: EncryptedMessage
-    ) => {
-      const states = (await messageDB.getAllEncryptionStates()).reduce(
-        (prev, curr) => {
-          return Object.assign(prev, { [curr.inboxId]: curr });
-        },
-        {} as { [key: string]: EncryptionState }
-      );
-      let found = states[message.inboxAddress];
-
-      if (
-        message.inboxAddress == keyset.deviceKeyset.inbox_keyset.inbox_address
-      ) {
-        try {
-          const envelope = Object.assign(
-            secureChannel.UnsealInitializationEnvelope(
-              keyset.deviceKeyset,
-              JSON.parse(message.encryptedContent)
-            ),
-            { timestamp: message.timestamp }
-          );
-          const session = await secureChannel.NewDoubleRatchetRecipientSession(
-            keyset.deviceKeyset,
-            envelope
-          );
-
-          let decryptedContent: Message | null = null;
-          let newState: any | null = null;
-
-          let conversationId =
-            session.user_address + '/' + session.user_address;
-
-          let updatedUserProfile: secureChannel.UserProfile | undefined;
-          decryptedContent = JSON.parse(session.message);
-
-          if (session.user_address == self_address) {
-            conversationId =
-              decryptedContent?.channelId + '/' + decryptedContent?.channelId;
-            session.user_address = decryptedContent!.channelId;
-          }
-          if (decryptedContent?.content?.type === 'delete-conversation') {
-            await deleteEncryptionStates({ conversationId });
-            await deleteInboxMessages(
-              keyset.deviceKeyset.inbox_keyset,
-              [envelope.timestamp],
-              apiClient
-            );
-            return;
-          }
-
-          const encryptionStates = await messageDB.getEncryptionStates({
-            conversationId,
-          });
-          const existing = encryptionStates.filter(
-            (e) => JSON.parse(e.state).tag == session.tag
-          );
-          for (const e of existing) {
-            await messageDB.deleteEncryptionState(e);
-          }
-
-          const inbox_key = await secureChannel.NewInboxKeyset();
-          newState = JSON.stringify({
-            ratchet_state: session.state,
-            receiving_inbox: inbox_key,
-            tag: session.tag,
-            sending_inbox: {
-              inbox_address: session.return_inbox_address,
-              inbox_encryption_key: session.return_inbox_encryption_key,
-              inbox_public_key: session.return_inbox_public_key,
-              inbox_private_key: session.return_inbox_private_key,
-            },
-          });
-          if (envelope.user_address != self_address) {
-            updatedUserProfile = {
-              user_address: envelope.user_address,
-              user_icon: envelope.user_icon,
-              display_name: envelope.display_name,
-            };
-          }
-          enqueueOutbound(async () => {
-            return [
-              JSON.stringify({
-                type: 'listen',
-                inbox_addresses: [inbox_key.inbox_address],
-              }),
-            ];
-          });
-
-          if (decryptedContent && newState) {
-            const newEncryptionState: EncryptionState = {
-              state: newState,
-              timestamp: message.timestamp,
-              inboxId: inbox_key.inbox_address,
-              conversationId: conversationId,
-            };
-
-            await messageDB.saveEncryptionState(newEncryptionState, true);
-            const conversation = await messageDB.getConversation({
-              conversationId,
-            });
-            await saveMessage(
-              decryptedContent,
-              messageDB,
-              session.user_address,
-              session.user_address,
-              'direct',
-              updatedUserProfile ?? {
-                user_icon:
-                  conversation?.conversation?.icon ??
-                  DefaultImages.UNKNOWN_USER,
-                display_name:
-                  conversation?.conversation?.displayName ?? t`Unknown User`,
-              }
-            );
-            await addMessage(
-              queryClient,
-              session.user_address,
-              session.user_address,
-              decryptedContent
-            );
-            addOrUpdateConversation(
-              queryClient,
-              session.user_address,
-              envelope.timestamp,
-              0,
-              updatedUserProfile ?? {
-                user_icon:
-                  conversation?.conversation?.icon ??
-                  DefaultImages.UNKNOWN_USER,
-                display_name:
-                  conversation?.conversation?.displayName ?? t`Unknown User`,
-              }
-            );
-          } else {
-            console.error(t`Failed to decrypt message with any known state`);
-          }
-          await deleteInboxMessages(
-            keyset.deviceKeyset.inbox_keyset,
-            [envelope.timestamp],
-            apiClient
-          );
-        } catch (error) {
-          await deleteInboxMessages(
-            keyset.deviceKeyset.inbox_keyset,
-            [message.timestamp],
-            apiClient
-          );
-          return;
-        }
-        return;
-      }
-
-      if (!found) {
-        await deleteInboxMessages(
-          keyset.deviceKeyset.inbox_keyset,
-          [message.timestamp],
-          apiClient
-        );
-        return;
-      }
-
-      const conversationId = found.conversationId;
-      const conversation = await messageDB.getConversation({ conversationId });
-
-      let decryptedContent: Message | null = null;
-      let newState: string | null = null;
-
-      let keys = JSON.parse(found.state);
-      let updatedUserProfile: secureChannel.UserProfile | undefined;
-      let sentAccept: boolean | undefined;
-      if (keys.sending_inbox) {
-        // secureChannel.DoubleRatchetStateAndInboxKeys
-        if (keys.sending_inbox.inbox_public_key === '') {
-          try {
-            const result =
-              await secureChannel.ConfirmDoubleRatchetSenderSession(
-                JSON.parse(found.state),
-                JSON.parse(message.encryptedContent)
-              );
-            decryptedContent = JSON.parse(result.message);
-            newState = JSON.stringify({
-              ratchet_state: result.ratchet_state,
-              receiving_inbox: result.receiving_inbox,
-              sending_inbox: result.sending_inbox,
-              tag: result.tag,
-            });
-            sentAccept = true;
-            if (result.user_profile.user_address != self_address) {
-              updatedUserProfile = result.user_profile;
-            }
-            if (decryptedContent?.content?.type === 'delete-conversation') {
-              await deleteEncryptionStates({ conversationId });
-              await deleteInboxMessages(
-                keys.receiving_inbox,
-                [message.timestamp],
-                apiClient
-              );
-              return;
-            }
-          } catch (error) {
-            await deleteInboxMessages(
-              keys.receiving_inbox,
-              [message.timestamp],
-              apiClient
-            );
-            await messageDB.deleteEncryptionState(found);
-            return;
-          }
-        } else {
-          try {
-            const result = await secureChannel.DoubleRatchetInboxDecrypt(
-              JSON.parse(found.state),
-              JSON.parse(message.encryptedContent)
-            );
-            const maybeInit = result as {
-              receiving_inbox: secureChannel.InboxKeyset;
-              user_profile: secureChannel.UserProfile;
-              tag: any;
-              sending_inbox: secureChannel.SendingInbox;
-              ratchet_state: string;
-              message: string;
-            };
-
-            if (maybeInit.user_profile) {
-              newState = JSON.stringify({
-                ratchet_state: maybeInit.ratchet_state,
-                receiving_inbox: maybeInit.receiving_inbox,
-                sending_inbox: maybeInit.sending_inbox,
-                tag: maybeInit.tag,
-              });
-            } else {
-              newState = JSON.stringify({
-                ratchet_state: result.ratchet_state,
-                receiving_inbox: keys.receiving_inbox,
-                sending_inbox: keys.sending_inbox,
-                tag: keys.tag,
-              });
-            }
-            decryptedContent = JSON.parse(result.message);
-            sentAccept = found.sentAccept;
-            if ((decryptedContent as any)?.content?.type === 'delete-conversation') {
-              await deleteEncryptionStates({ conversationId });
-              await deleteInboxMessages(
-                keys.receiving_inbox,
-                [message.timestamp],
-                apiClient
-              );
-              return;
-            }
-          } catch (error) {
-            await deleteInboxMessages(
-              keys.receiving_inbox,
-              [message.timestamp],
-              apiClient
-            );
-            await messageDB.deleteEncryptionState(found);
-            return;
-          }
-        }
-      } else {
-        try {
-          const hub_key = await messageDB.getSpaceKey(
-            conversationId.split('/')[0],
-            'hub'
-          );
-          const result = Buffer.from(
-            new Uint8Array(
-              await secureChannel.UnsealHubEnvelope(
-                {
-                  type: 'ed448',
-                  public_key: [
-                    ...new Uint8Array(Buffer.from(hub_key.publicKey, 'hex')),
-                  ],
-                  private_key: [
-                    ...new Uint8Array(Buffer.from(hub_key.privateKey, 'hex')),
-                  ],
-                },
-                JSON.parse(message.encryptedContent)
-              )
-            )
-          ).toString('utf-8');
-          const envelope = JSON.parse(result);
-          if (envelope.type === 'message') {
-            const decrypted = JSON.parse(
-              await secureChannel.TripleRatchetDecrypt(
-                JSON.stringify({
-                  ratchet_state: keys.state,
-                  envelope: JSON.stringify(envelope.message),
-                })
-              )
-            );
-            const output = Buffer.from(
-              new Uint8Array(decrypted.message)
-            ).toString('utf-8');
-            decryptedContent = JSON.parse(output);
-
-            if (decryptedContent) {
-              const space = await messageDB.getSpace(
-                conversationId.split('/')[0]
-              );
-
-              // enforce non-repudiability
-              if (
-                space &&
-                !space.isRepudiable &&
-                decryptedContent.publicKey &&
-                decryptedContent.signature
-              ) {
-                const participant = await messageDB.getSpaceMember(
-                  space.spaceId,
-                  decryptedContent.content.senderId
-                );
-                const sh = await sha256.digest(
-                  Buffer.from(decryptedContent.publicKey, 'hex')
-                );
-                const inboxAddress = base58btc.baseEncode(sh.bytes);
-                const messageId = await crypto.subtle.digest(
-                  'SHA-256',
-                  Buffer.from(
-                    decryptedContent.nonce +
-                      'post' +
-                      decryptedContent.content.senderId +
-                      canonicalize(decryptedContent.content as any),
-                    'utf-8'
-                  )
-                );
-                const inboxMismatch =
-                  participant.inbox_address !== inboxAddress &&
-                  participant.inbox_address;
-                const messageIdMismatch =
-                  decryptedContent.messageId !==
-                  Buffer.from(messageId).toString('hex');
-
-                if (inboxMismatch || messageIdMismatch) {
-                  console.warn(t`invalid address for signature`);
-                  decryptedContent.publicKey = undefined;
-                  decryptedContent.signature = undefined;
-                } else {
-                  if (
-                    ch.js_verify_ed448(
-                      Buffer.from(decryptedContent.publicKey, 'hex').toString(
-                        'base64'
-                      ),
-                      Buffer.from(messageId).toString('base64'),
-                      Buffer.from(decryptedContent.signature, 'hex').toString(
-                        'base64'
-                      )
-                    ) !== 'true'
-                  ) {
-                    console.warn('invalid signature');
-                    decryptedContent.publicKey = undefined;
-                    decryptedContent.signature = undefined;
-                  }
-                }
-              }
-
-              if (
-                decryptedContent?.content.type === 'update-profile' &&
-                (!decryptedContent?.publicKey || !decryptedContent?.signature)
-              ) {
-                decryptedContent = null;
-              }
-            }
-          } else if (envelope.type === 'control') {
-            const exteriorEnvelope = JSON.parse(message.encryptedContent);
-            if (envelope.message.type === 'join') {
-              const participant = envelope.message.participant;
-              const pointResult = ch.js_verify_point(
-                JSON.stringify({
-                  ratchet_state: keys.state,
-                  point: participant.pubKey,
-                  index: participant.id,
-                })
-              );
-              if (pointResult === 'true') {
-                const msg = Buffer.from(
-                  participant.address +
-                    participant.id +
-                    participant.inboxAddress +
-                    participant.pubKey +
-                    participant.inboxKey +
-                    participant.identityKey +
-                    participant.preKey +
-                    participant.userIcon +
-                    participant.displayName,
-                  'utf-8'
-                ).toString('base64');
-                const result = ch.js_verify_ed448(
-                  Buffer.from(participant.inboxPubKey, 'hex').toString(
-                    'base64'
-                  ),
-                  msg,
-                  participant.signature
-                );
-                if (result === 'true') {
-                  messageDB.saveSpaceMember(conversationId.split('/')[0], {
-                    user_address: participant.address,
-                    user_icon: participant.userIcon,
-                    display_name: participant.displayName,
-                    inbox_address: participant.inboxAddress,
-                    isKicked: false,
-                  });
-                  await queryClient.setQueryData(
-                    buildSpaceMembersKey({
-                      spaceId: conversationId.split('/')[0],
-                    }),
-                    (oldData: secureChannel.UserProfile[]) => {
-                      return [
-                        ...(oldData ?? []),
-                        {
-                          user_address: participant.address,
-                          user_icon: participant.userIcon,
-                          display_name: participant.displayName,
-                          // isKicked intentionally omitted here (defaults to false on fetch)
-                        },
-                      ];
-                    }
-                  );
-                  const ratchet = JSON.parse(keys.state);
-                  ratchet.id_peer_map = {
-                    ...ratchet.id_peer_map,
-                    [participant.id]: {
-                      public_key: Buffer.from(
-                        participant.inboxKey,
-                        'hex'
-                      ).toString('base64'),
-                      identity_public_key: Buffer.from(
-                        participant.identityKey,
-                        'hex'
-                      ).toString('base64'),
-                      signed_pre_public_key: Buffer.from(
-                        participant.preKey,
-                        'hex'
-                      ).toString('base64'),
-                    },
-                  };
-                  ratchet.peer_id_map = {
-                    ...ratchet.peer_id_map,
-                    [Buffer.from(participant.inboxKey, 'hex').toString(
-                      'base64'
-                    )]: participant.id,
-                  };
-                  newState = JSON.stringify({
-                    ...keys,
-                    state: JSON.stringify(ratchet),
-                  });
-                  const space = await messageDB.getSpace(
-                    conversationId.split('/')[0]
-                  );
-                  const messageId = await crypto.subtle.digest(
-                    'SHA-256',
-                    Buffer.from('join' + participant.inboxAddress, 'utf-8')
-                  );
-                  const msg = {
-                    channelId: space!.defaultChannelId,
-                    spaceId: conversationId.split('/')[0],
-                    messageId: Buffer.from(messageId).toString('hex'),
-                    digestAlgorithm: 'SHA-256',
-                    nonce: Buffer.from(messageId).toString('hex'),
-                    createdDate: Date.now(),
-                    modifiedDate: Date.now(),
-                    lastModifiedHash: '',
-                    content: {
-                      senderId: participant.address,
-                      type: 'join',
-                    } as JoinMessage,
-                  } as Message;
-                  await saveMessage(
-                    msg,
-                    messageDB,
-                    conversationId.split('/')[0],
-                    space!.defaultChannelId,
-                    'group',
-                    {}
-                  );
-                  await addMessage(
-                    queryClient,
-                    conversationId.split('/')[0],
-                    space!.defaultChannelId,
-                    msg
-                  );
-                }
-              } else {
-                console.error(pointResult);
-              }
-            } else if (envelope.message.type === 'sync-peer-map') {
-              let reg = spaceInfo.current[conversationId.split('/')[0]];
-              if (!reg) {
-                reg = (await apiClient.getSpace(conversationId.split('/')[0]))
-                  .data;
-                spaceInfo.current[conversationId.split('/')[0]] = reg;
-              }
-
-              if (
-                reg.owner_public_keys.includes(
-                  exteriorEnvelope.owner_public_key
-                ) ||
-                syncInfo.current[conversationId.split('/')[0]]
-              ) {
-                const verify = JSON.parse(
-                  ch.js_verify_ed448(
-                    Buffer.from(
-                      exteriorEnvelope.owner_public_key,
-                      'hex'
-                    ).toString('base64'),
-                    Buffer.from(exteriorEnvelope.envelope, 'utf-8').toString(
-                      'base64'
-                    ),
-                    Buffer.from(
-                      exteriorEnvelope.owner_signature,
-                      'hex'
-                    ).toString('base64')
-                  )
-                );
-                if (verify) {
-                  const ratchet = JSON.parse(keys.state);
-                  ratchet.id_peer_map = envelope.message.peerMap.id_peer_map;
-                  ratchet.peer_id_map = envelope.message.peerMap.peer_id_map;
-                  newState = JSON.stringify({
-                    ...keys,
-                    state: JSON.stringify(ratchet),
-                  });
-                }
-              }
-            } else if (envelope.message.type === 'space-manifest') {
-              let reg = spaceInfo.current[conversationId.split('/')[0]];
-              if (!reg) {
-                reg = (await apiClient.getSpace(conversationId.split('/')[0]))
-                  .data;
-                spaceInfo.current[conversationId.split('/')[0]] = reg;
-              }
-              const manifest = envelope.message
-                .manifest as secureChannel.SpaceManifest;
-              if (reg.owner_public_keys.includes(manifest.owner_public_key)) {
-                const verify = JSON.parse(
-                  ch.js_verify_ed448(
-                    Buffer.from(manifest.owner_public_key, 'hex').toString(
-                      'base64'
-                    ),
-                    Buffer.from(
-                      new Uint8Array([
-                        ...new Uint8Array(
-                          Buffer.from(manifest.space_manifest, 'utf-8')
-                        ),
-                        ...int64ToBytes(manifest.timestamp),
-                      ])
-                    ).toString('base64'),
-                    Buffer.from(manifest.owner_signature, 'hex').toString(
-                      'base64'
-                    )
-                  )
-                );
-                if (verify) {
-                  const ciphertext = JSON.parse(manifest.space_manifest) as {
-                    ciphertext: string;
-                    initialization_vector: string;
-                    associated_data: string;
-                  };
-                  const config_key = await messageDB.getSpaceKey(
-                    conversationId.split('/')[0],
-                    'config'
-                  );
-                  const space = JSON.parse(
-                    Buffer.from(
-                      JSON.parse(
-                        ch.js_decrypt_inbox_message(
-                          JSON.stringify({
-                            inbox_private_key: [
-                              ...new Uint8Array(
-                                Buffer.from(config_key.privateKey, 'hex')
-                              ),
-                            ],
-                            ephemeral_public_key: [
-                              ...new Uint8Array(
-                                Buffer.from(
-                                  manifest.ephemeral_public_key,
-                                  'hex'
-                                )
-                              ),
-                            ],
-                            ciphertext: ciphertext,
-                          })
-                        )
-                      )
-                    ).toString('utf-8')
-                  ) as Space;
-                  await messageDB.saveSpace(space);
-                  await queryClient.setQueryData(
-                    buildSpaceKey({ spaceId: conversationId.split('/')[0] }),
-                    () => {
-                      return space;
-                    }
-                  );
-                }
-              }
-            } else if (envelope.message.type === 'leave') {
-              const hubKey = await messageDB.getSpaceKey(
-                conversationId.split('/')[0],
-                'hub'
-              );
-
-              const verify = JSON.parse(
-                ch.js_verify_ed448(
-                  Buffer.from(envelope.message.inboxPublicKey, 'hex').toString(
-                    'base64'
-                  ),
-                  Buffer.from(
-                    new Uint8Array([
-                      ...new Uint8Array(
-                        Buffer.from('delete' + hubKey.publicKey, 'utf-8')
-                      ),
-                    ])
-                  ).toString('base64'),
-                  Buffer.from(envelope.message.inboxSignature, 'hex').toString(
-                    'base64'
-                  )
-                )
-              );
-              const sh = await sha256.digest(
-                Buffer.from(envelope.message.inboxPublicKey, 'hex')
-              );
-              const inboxAddress = base58btc.baseEncode(sh.bytes);
-              if (verify) {
-                const members = await messageDB.getSpaceMembers(
-                  conversationId.split('/')[0]
-                );
-                for (const member of members) {
-                  if (member.inbox_address == inboxAddress) {
-                    await messageDB.saveSpaceMember(
-                      conversationId.split('/')[0],
-                      { ...member, inbox_address: '' }
-                    );
-                    await queryClient.setQueryData(
-                      buildSpaceMembersKey({
-                        spaceId: conversationId.split('/')[0],
-                      }),
-                      (
-                        oldData: (secureChannel.UserProfile & {
-                          inbox_address: string;
-                          isKicked?: boolean;
-                        })[]
-                      ) => {
-                        const previous = oldData ?? [];
-                        return previous.map((m) =>
-                          m.user_address === member.user_address
-                            ? { ...m, inbox_address: '' }
-                            : m
-                        );
-                      }
-                    );
-                    const space = await messageDB.getSpace(
-                      conversationId.split('/')[0]
-                    );
-                    const messageId = await crypto.subtle.digest(
-                      'SHA-256',
-                      Buffer.from('leave' + member.inbox_address, 'utf-8')
-                    );
-                    const msg = {
-                      channelId: space!.defaultChannelId,
-                      spaceId: conversationId.split('/')[0],
-                      messageId: Buffer.from(messageId).toString('hex'),
-                      digestAlgorithm: 'SHA-256',
-                      nonce: Buffer.from(messageId).toString('hex'),
-                      createdDate: Date.now(),
-                      modifiedDate: Date.now(),
-                      lastModifiedHash: '',
-                      content: {
-                        senderId: member.user_address,
-                        type: 'leave',
-                      } as LeaveMessage,
-                    } as Message;
-                    await saveMessage(
-                      msg,
-                      messageDB,
-                      conversationId.split('/')[0],
-                      space!.defaultChannelId,
-                      'group',
-                      {}
-                    );
-                    await addMessage(
-                      queryClient,
-                      conversationId.split('/')[0],
-                      space!.defaultChannelId,
-                      msg
-                    );
-                    break;
-                  }
-                }
-              }
-            } else if (envelope.message.type === 'rekey') {
-              let reg = spaceInfo.current[conversationId.split('/')[0]];
-              if (!reg) {
-                reg = (await apiClient.getSpace(conversationId.split('/')[0]))
-                  .data;
-                spaceInfo.current[conversationId.split('/')[0]] = reg;
-              }
-              if (
-                reg.owner_public_keys.includes(
-                  exteriorEnvelope.owner_public_key
-                )
-              ) {
-                const verify = JSON.parse(
-                  ch.js_verify_ed448(
-                    Buffer.from(
-                      exteriorEnvelope.owner_public_key,
-                      'hex'
-                    ).toString('base64'),
-                    Buffer.from(exteriorEnvelope.envelope, 'utf-8').toString(
-                      'base64'
-                    ),
-                    Buffer.from(
-                      exteriorEnvelope.owner_signature,
-                      'hex'
-                    ).toString('base64')
-                  )
-                );
-                if (verify) {
-                  const info = JSON.parse(envelope.message.info);
-                  const inner_envelope = JSON.parse(
-                    Buffer.from(
-                      new Uint8Array(
-                        await secureChannel.UnsealInboxEnvelope(
-                          keyset.deviceKeyset.inbox_keyset.inbox_encryption_key
-                            .private_key,
-                          info
-                        )
-                      )
-                    ).toString('utf-8')
-                  );
-                  const configPub = Buffer.from(
-                    JSON.parse(
-                      ch.js_get_pubkey_x448(
-                        Buffer.from(inner_envelope.configKey, 'hex').toString(
-                          'base64'
-                        )
-                      )
-                    ),
-                    'base64'
-                  ).toString('hex');
-                  await messageDB.saveSpaceKey({
-                    spaceId: conversationId.split('/')[0],
-                    keyId: 'config',
-                    privateKey: inner_envelope.configKey,
-                    publicKey: configPub,
-                  });
-                  let template = JSON.parse(inner_envelope.state);
-                  template.peer_key = Buffer.from(
-                    new Uint8Array(
-                      keyset.deviceKeyset.inbox_keyset.inbox_encryption_key.private_key
-                    )
-                  ).toString('base64');
-                  newState = JSON.stringify({
-                    ...keys,
-                    state: JSON.stringify(template),
-                  });
-                  const space = await messageDB.getSpace(
-                    conversationId.split('/')[0]
-                  );
-                  if (envelope.message.kick) {
-                    const messageId = await crypto.subtle.digest(
-                      'SHA-256',
-                      Buffer.from('kick' + envelope.message.kick, 'utf-8')
-                    );
-                    const msg = {
-                      channelId: space!.defaultChannelId,
-                      spaceId: conversationId.split('/')[0],
-                      messageId: Buffer.from(messageId).toString('hex'),
-                      digestAlgorithm: 'SHA-256',
-                      nonce: Buffer.from(messageId).toString('hex'),
-                      createdDate: Date.now(),
-                      modifiedDate: Date.now(),
-                      lastModifiedHash: '',
-                      content: {
-                        senderId: envelope.message.kick,
-                        type: 'kick',
-                      } as KickMessage,
-                    } as Message;
-                    await saveMessage(
-                      msg,
-                      messageDB,
-                      conversationId.split('/')[0],
-                      space!.defaultChannelId,
-                      'group',
-                      {}
-                    );
-                    await addMessage(
-                      queryClient,
-                      conversationId.split('/')[0],
-                      space!.defaultChannelId,
-                      msg
-                    );
-                  }
-
-                  if (space?.inviteUrl) {
-                    space.inviteUrl = `${getInviteUrlBase(true)}#spaceId=${space.spaceId}&configKey=${inner_envelope.configKey}`;
-                    await messageDB.saveSpace(space);
-                  }
-                }
-              }
-            } else if (envelope.message.type === 'kick') {
-              let reg = spaceInfo.current[conversationId.split('/')[0]];
-              if (!reg) {
-                reg = (await apiClient.getSpace(conversationId.split('/')[0]))
-                  .data;
-                spaceInfo.current[conversationId.split('/')[0]] = reg;
-              }
-              if (
-                reg.owner_public_keys.includes(
-                  exteriorEnvelope.owner_public_key
-                )
-              ) {
-                const verify = JSON.parse(
-                  ch.js_verify_ed448(
-                    Buffer.from(
-                      exteriorEnvelope.owner_public_key,
-                      'hex'
-                    ).toString('base64'),
-                    Buffer.from(exteriorEnvelope.envelope, 'utf-8').toString(
-                      'base64'
-                    ),
-                    Buffer.from(
-                      exteriorEnvelope.owner_signature,
-                      'hex'
-                    ).toString('base64')
-                  )
-                );
-                if (verify) {
-                  if (envelope.message.kick === self_address) {
-                    const spaceId = conversationId.split('/')[0];
-                    try {
-                      const space = await messageDB.getSpace(spaceId);
-                      if (typeof window !== 'undefined' && (window as any).dispatchEvent) {
-                        (window as any).dispatchEvent(
-                          new CustomEvent('quorum:kick-toast', {
-                            detail: { spaceName: space?.spaceName || spaceId },
-                          })
-                        );
-                      }
-                    } catch {}
-                    // Immediately navigate away from the space view when kicked
-                    navigate('/messages', { replace: true, state: { from: 'kicked', spaceId } });
-                    const hubKey = await messageDB.getSpaceKey(spaceId, 'hub');
-                    const inboxKey = await messageDB.getSpaceKey(
-                      spaceId,
-                      'inbox'
-                    );
-                    await apiClient.postHubDelete({
-                      hub_address: hubKey.address!,
-                      hub_public_key: hubKey.publicKey,
-                      hub_signature: Buffer.from(
-                        JSON.parse(
-                          ch.js_sign_ed448(
-                            Buffer.from(hubKey.privateKey, 'hex').toString(
-                              'base64'
-                            ),
-                            Buffer.from(
-                              new Uint8Array([
-                                ...new Uint8Array(
-                                  Buffer.from(
-                                    'delete' + inboxKey.publicKey,
-                                    'utf-8'
-                                  )
-                                ),
-                              ])
-                            ).toString('base64')
-                          )
-                        ),
-                        'base64'
-                      ).toString('hex'),
-                      inbox_public_key: inboxKey.publicKey,
-                      inbox_signature: Buffer.from(
-                        JSON.parse(
-                          ch.js_sign_ed448(
-                            Buffer.from(inboxKey.privateKey, 'hex').toString(
-                              'base64'
-                            ),
-                            Buffer.from(
-                              new Uint8Array([
-                                ...new Uint8Array(
-                                  Buffer.from(
-                                    'delete' + hubKey.publicKey,
-                                    'utf-8'
-                                  )
-                                ),
-                              ])
-                            ).toString('base64')
-                          )
-                        ),
-                        'base64'
-                      ).toString('hex'),
-                    });
-                    const states = await messageDB.getEncryptionStates({
-                      conversationId: spaceId + '/' + spaceId,
-                    });
-                    for (const state of states) {
-                      await messageDB.deleteEncryptionState(state);
-                    }
-                    const messages = await messageDB.getAllSpaceMessages({
-                      spaceId,
-                    });
-                    for (const message of messages) {
-                      await messageDB.deleteMessage(message.messageId);
-                    }
-                    const members = await messageDB.getSpaceMembers(spaceId);
-                    for (const member of members) {
-                      await messageDB.deleteSpaceMember(
-                        spaceId,
-                        member.user_address
-                      );
-                    }
-                    const keys = await messageDB.getSpaceKeys(spaceId);
-                    for (const key of keys) {
-                      await messageDB.deleteSpaceKey(spaceId, key.keyId);
-                    }
-                    let userConfig = await messageDB.getUserConfig({
-                      address: self_address,
-                    });
-                    userConfig = {
-                      ...(userConfig ?? { address: self_address }),
-                      spaceIds: [
-                        ...(userConfig?.spaceIds.filter((s) => s != spaceId) ??
-                          []),
-                      ],
-                    };
-                    await saveConfig({ config: userConfig, keyset });
-                    await queryClient.setQueryData(
-                      buildConfigKey({ userAddress: self_address }),
-                      () => userConfig
-                    );
-                    await messageDB.deleteSpace(spaceId);
-                    return;
-                  }
-                  // If someone else was kicked, mark them inactive locally
-                  if (
-                    envelope.message.kick &&
-                    envelope.message.kick !== self_address
-                  ) {
-                    const spaceId = conversationId.split('/')[0];
-                    const kickedAddress = envelope.message.kick;
-                    const kicked = await messageDB.getSpaceMember(
-                      spaceId,
-                      kickedAddress
-                    );
-                    if (kicked) {
-                      await messageDB.saveSpaceMember(spaceId, {
-                        ...kicked,
-                        inbox_address: '',
-                      });
-                      await queryClient.setQueryData(
-                        buildSpaceMembersKey({ spaceId }),
-                        (
-                          oldData: (secureChannel.UserProfile & {
-                            inbox_address: string;
-                          })[]
-                        ) => {
-                          const previous = oldData ?? [];
-                          return previous.map((m) =>
-                            m.user_address === kickedAddress
-                              ? { ...m, inbox_address: '' }
-                              : m
-                          );
-                        }
-                      );
-                    }
-                  }
-                }
-              }
-            } else if (envelope.message.type === 'sync') {
-              await synchronizeAll(
-                conversationId.split('/')[0],
-                envelope.message.inboxAddress
-              );
-            } else if (envelope.message.type === 'sync-request') {
-              if (envelope.message.expiry > Date.now()) {
-                await informSyncData(
-                  conversationId.split('/')[0],
-                  envelope.message.inboxAddress,
-                  envelope.message.messageCount,
-                  envelope.message.memberCount
-                );
-              }
-            } else if (envelope.message.type === 'sync-info') {
-              if (
-                syncInfo.current[conversationId.split('/')[0]] &&
-                syncInfo.current[conversationId.split('/')[0]].expiry >
-                  Date.now()
-              ) {
-                if (
-                  envelope.message.inboxAddress &&
-                  envelope.message.messageCount &&
-                  envelope.message.memberCount
-                ) {
-                  syncInfo.current[
-                    conversationId.split('/')[0]
-                  ].candidates.push(envelope.message);
-                  // reset the timeout to be 1s to more aggressively grab viable candidates for sync instead of waiting the full 30s
-                  clearTimeout(
-                    syncInfo.current[conversationId.split('/')[0]].invokable
-                  );
-                  syncInfo.current[conversationId.split('/')[0]].invokable =
-                    setTimeout(
-                      () => initiateSync(conversationId.split('/')[0]),
-                      1000
-                    );
-                }
-              }
-            } else if (envelope.message.type === 'sync-initiate') {
-              if (envelope.message.inboxAddress) {
-                await directSync(
-                  conversationId.split('/')[0],
-                  envelope.message
-                );
-              }
-            } else if (envelope.message.type === 'sync-members') {
-              let reg = spaceInfo.current[conversationId.split('/')[0]];
-              if (!reg) {
-                reg = (await apiClient.getSpace(conversationId.split('/')[0]))
-                  .data;
-                spaceInfo.current[conversationId.split('/')[0]] = reg;
-              }
-
-              if (
-                reg.owner_public_keys.includes(
-                  exteriorEnvelope.owner_public_key
-                ) ||
-                syncInfo.current[conversationId.split('/')[0]]
-              ) {
-                const verify = JSON.parse(
-                  ch.js_verify_ed448(
-                    Buffer.from(
-                      exteriorEnvelope.owner_public_key,
-                      'hex'
-                    ).toString('base64'),
-                    Buffer.from(exteriorEnvelope.envelope, 'utf-8').toString(
-                      'base64'
-                    ),
-                    Buffer.from(
-                      exteriorEnvelope.owner_signature,
-                      'hex'
-                    ).toString('base64')
-                  )
-                );
-                if (verify) {
-                  for (const member of envelope.message.members) {
-                    try {
-                      const existing = await messageDB.getSpaceMember(
-                        conversationId.split('/')[0],
-                        (member as any).user_address
-                      );
-                      await messageDB.saveSpaceMember(
-                        conversationId.split('/')[0],
-                        {
-                          ...(member as any),
-                          isKicked: existing?.isKicked ?? false,
-                        } as any
-                      );
-                    } catch {
-                      await messageDB.saveSpaceMember(
-                        conversationId.split('/')[0],
-                        member as any
-                      );
-                    }
-                  }
-                  await queryClient.setQueryData(
-                    buildSpaceMembersKey({
-                      spaceId: conversationId.split('/')[0],
-                    }),
-                    (oldData: (secureChannel.UserProfile & { isKicked?: boolean })[]) => {
-                      const existingMap = new Map(
-                        (oldData ?? []).map((m) => [m.user_address, m])
-                      );
-                      const merged = (envelope.message.members as any[]).map((m) => {
-                        const prev = existingMap.get(m.user_address);
-                        return { ...m, isKicked: prev?.isKicked ?? false };
-                      });
-                      return [...(oldData ?? []), ...merged];
-                    }
-                  );
-                }
-              }
-            } else if (envelope.message.type === 'verify-kicked') {
-              if (Array.isArray(envelope.message.addresses)) {
-                const spaceId = conversationId.split('/')[0];
-                for (const address of envelope.message.addresses) {
-                  const member = await messageDB.getSpaceMember(spaceId, address);
-                  if (member) {
-                    await messageDB.saveSpaceMember(spaceId, {
-                      ...member,
-                      isKicked: true,
-                    });
-                  }
-                }
-                await queryClient.setQueryData(
-                  buildSpaceMembersKey({ spaceId }),
-                  (
-                    oldData: (secureChannel.UserProfile & {
-                      inbox_address: string;
-                      isKicked?: boolean;
-                    })[]
-                  ) => {
-                    const previous = oldData ?? [];
-                    const mark = new Set(envelope.message.addresses as string[]);
-                    return previous.map((m) =>
-                      mark.has(m.user_address) ? { ...m, isKicked: true } : m
-                    );
-                  }
-                );
-              }
-            } else if (envelope.message.type === 'sync-messages') {
-              let reg = spaceInfo.current[conversationId.split('/')[0]];
-              if (!reg) {
-                reg = (await apiClient.getSpace(conversationId.split('/')[0]))
-                  .data;
-                spaceInfo.current[conversationId.split('/')[0]] = reg;
-              }
-
-              if (
-                reg.owner_public_keys.includes(
-                  exteriorEnvelope.owner_public_key
-                ) ||
-                syncInfo.current[conversationId.split('/')[0]]
-              ) {
-                const verify = JSON.parse(
-                  ch.js_verify_ed448(
-                    Buffer.from(
-                      exteriorEnvelope.owner_public_key,
-                      'hex'
-                    ).toString('base64'),
-                    Buffer.from(exteriorEnvelope.envelope, 'utf-8').toString(
-                      'base64'
-                    ),
-                    Buffer.from(
-                      exteriorEnvelope.owner_signature,
-                      'hex'
-                    ).toString('base64')
-                  )
-                );
-                if (verify) {
-                  const space = await messageDB.getSpace(
-                    conversationId.split('/')[0]
-                  );
-                  for (const message of envelope.message.messages) {
-                    // enforce non-repudiability
-                    if (
-                      space &&
-                      !space.isRepudiable &&
-                      message.publicKey &&
-                      message.signature
-                    ) {
-                      const participant = await messageDB.getSpaceMember(
-                        space.spaceId,
-                        message.content.senderId
-                      );
-                      const sh = await sha256.digest(
-                        Buffer.from(message.publicKey, 'hex')
-                      );
-                      const inboxAddress = base58btc.baseEncode(sh.bytes);
-                      const messageId = await crypto.subtle.digest(
-                        'SHA-256',
-                        Buffer.from(
-                          message.nonce +
-                            'post' +
-                            message.content.senderId +
-                            canonicalize(message.content as any),
-                          'utf-8'
-                        )
-                      );
-                      if (
-                        (participant.inbox_address !== inboxAddress &&
-                          participant.inbox_address) ||
-                        message.messageId !==
-                          Buffer.from(messageId).toString('hex')
-                      ) {
-                        message.publicKey = undefined;
-                        message.signature = undefined;
-                      } else {
-                        if (
-                          ch.js_verify_ed448(
-                            Buffer.from(message.publicKey, 'hex').toString(
-                              'base64'
-                            ),
-                            Buffer.from(messageId).toString('base64'),
-                            Buffer.from(message.signature, 'hex').toString(
-                              'base64'
-                            )
-                          ) !== 'true'
-                        ) {
-                          console.warn('invalid signature');
-                          message.publicKey = undefined;
-                          message.signature = undefined;
-                        }
-                      }
-                    }
-                    await saveMessage(
-                      message,
-                      messageDB,
-                      conversationId.split('/')[0],
-                      message.channelId,
-                      'group',
-                      {}
-                    );
-                  }
-                  const channelIds = envelope.message.messages
-                    .map((m: any) => m.channelId)
-                    .sort();
-                  const checked = {} as { [id: string]: boolean };
-                  for (const channelId of channelIds) {
-                    if (!checked[channelId]) {
-                      checked[channelId] = true;
-                      queryClient.refetchQueries({
-                        queryKey: buildMessagesKey({
-                          spaceId: conversationId.split('/')[0],
-                          channelId: channelId,
-                        }),
-                      });
-                    }
-                  }
-                }
-              }
-            }
-          }
-        } catch {}
-      }
-
-      if (newState) {
-        const newEncryptionState: EncryptionState = {
-          state: newState,
-          timestamp: message.timestamp,
-          inboxId: found.inboxId,
-          sentAccept: sentAccept,
-          conversationId: conversationId,
-        };
-        await messageDB.saveEncryptionState(newEncryptionState, true);
-      }
-
-      if (decryptedContent) {
-        if (keys.sending_inbox) {
-          await saveMessage(
-            decryptedContent,
-            messageDB,
-            conversationId.split('/')[0],
-            conversationId.split('/')[0],
-            keys.sending_inbox ? 'direct' : 'group',
-            updatedUserProfile ?? {
-              user_icon: conversation.conversation?.icon,
-              display_name: conversation.conversation?.displayName,
-            }
-          );
-          await addMessage(
-            queryClient,
-            conversationId.split('/')[0],
-            conversationId.split('/')[0],
-            decryptedContent
-          );
-          addOrUpdateConversation(
-            queryClient,
-            conversationId.split('/')[0],
-            message.timestamp,
-            conversation.conversation?.lastReadTimestamp ?? 0,
-            updatedUserProfile ?? {
-              user_icon: conversation.conversation?.icon,
-              display_name: conversation.conversation?.displayName,
-            }
-          );
-        } else {
-          await saveMessage(
-            decryptedContent,
-            messageDB,
-            conversationId.split('/')[0],
-            decryptedContent.channelId,
-            keys.sending_inbox ? 'direct' : 'group',
-            updatedUserProfile ?? {
-              user_icon: conversation.conversation?.icon,
-              display_name: conversation.conversation?.displayName,
-            }
-          );
-          await addMessage(
-            queryClient,
-            conversationId.split('/')[0],
-            decryptedContent.channelId,
-            decryptedContent
-          );
-        }
-      }
-
-      if (keys.sending_inbox) {
-        await deleteInboxMessages(
-          keys.receiving_inbox,
-          [message.timestamp],
-          apiClient
-        );
-      } else {
-        const inbox_key = await messageDB.getSpaceKey(
-          conversationId.split('/')[0],
-          'inbox'
-        );
-        await deleteInboxMessages(
-          {
-            inbox_address: inbox_key.address!,
-            inbox_encryption_key: {} as never,
-            inbox_key: {
-              type: 'ed448',
-              public_key: [
-                ...new Uint8Array(Buffer.from(inbox_key.publicKey, 'hex')),
-              ],
-              private_key: [
-                ...new Uint8Array(Buffer.from(inbox_key.privateKey, 'hex')),
-              ],
-            },
-          },
-          [message.timestamp],
-          apiClient
-        );
-      }
-    },
-    [queryClient]
-  );
+  // MessageService will be instantiated after all dependencies are declared
+  // handleNewMessage will be defined after messageService instantiation
 
   const synchronizeAll = React.useCallback(
     async (spaceId: string, inboxAddress: string) => {
@@ -2859,208 +1066,7 @@ const MessageDBProvider: FC<MessageDBContextProps> = ({ children }) => {
     []
   );
 
-  const submitMessage = React.useCallback(
-    async (
-      address: string,
-      pendingMessage: string | object,
-      self: secureChannel.UserRegistration,
-      counterparty: secureChannel.UserRegistration,
-      queryClient: QueryClient,
-      currentPasskeyInfo: {
-        credentialId: string;
-        address: string;
-        publicKey: string;
-        displayName?: string;
-        pfpUrl?: string;
-        completedOnboarding: boolean;
-      },
-      keyset: {
-        deviceKeyset: secureChannel.DeviceKeyset;
-        userKeyset: secureChannel.UserKeyset;
-      },
-      inReplyTo?: string,
-      skipSigning?: boolean
-    ) => {
-      enqueueOutbound(async () => {
-        let outbounds: string[] = [];
-        const nonce = crypto.randomUUID();
-        const messageId = await crypto.subtle.digest(
-          'SHA-256',
-          Buffer.from(
-            nonce +
-              'post' +
-              currentPasskeyInfo.address +
-              (typeof pendingMessage === 'string'
-                ? pendingMessage
-                : JSON.stringify(pendingMessage)),
-            'utf-8'
-          )
-        );
-        const message = {
-          channelId: address!,
-          spaceId: address!,
-          messageId: Buffer.from(messageId).toString('hex'),
-          digestAlgorithm: 'SHA-256',
-          nonce: nonce,
-          createdDate: Date.now(),
-          modifiedDate: Date.now(),
-          lastModifiedHash: '',
-          content:
-            typeof pendingMessage === 'string'
-              ? ({
-                  type: 'post',
-                  senderId: currentPasskeyInfo.address,
-                  text: pendingMessage,
-                  repliesToMessageId: inReplyTo,
-                } as PostMessage)
-              : {
-                  ...(pendingMessage as any),
-                  senderId: currentPasskeyInfo.address,
-                },
-        } as Message;
-        let conversationId = address + '/' + address;
-        const conversation = await messageDB.getConversation({
-          conversationId,
-        });
-        let response = await messageDB.getEncryptionStates({ conversationId });
-        const inboxes = self.device_registrations
-          .map((d) => d.inbox_registration.inbox_address)
-          .concat(
-            counterparty.device_registrations.map(
-              (d) => d.inbox_registration.inbox_address
-            )
-          )
-          .sort();
-        for (const res of response) {
-          if (!inboxes.includes(JSON.parse(res.state).tag)) {
-            await messageDB.deleteEncryptionState(res);
-          }
-        }
-
-        response = await messageDB.getEncryptionStates({ conversationId });
-        let sets = response.map((e) => JSON.parse(e.state));
-
-        let sessions: secureChannel.SealedMessageAndMetadata[] = [];
-        // Sign DM unless explicitly skipped
-        if (!skipSigning) {
-          try {
-            const sig = ch.js_sign_ed448(
-              Buffer.from(
-                new Uint8Array(keyset.userKeyset.user_key.private_key)
-              ).toString('base64'),
-              Buffer.from(messageId).toString('base64')
-            );
-            message.publicKey = Buffer.from(
-              new Uint8Array(keyset.userKeyset.user_key.public_key)
-            ).toString('hex');
-            message.signature = Buffer.from(JSON.parse(sig), 'base64').toString(
-              'hex'
-            );
-          } catch {}
-        }
-
-        for (const inbox of inboxes.filter(
-          (i) => i !== keyset.deviceKeyset.inbox_keyset.inbox_address
-        )) {
-          const set = sets.find((s) => s.tag === inbox);
-          if (set) {
-            if (set.sending_inbox.inbox_public_key === '') {
-              sessions = [
-                ...sessions,
-                ...secureChannel.DoubleRatchetInboxEncryptForceSenderInit(
-                  keyset.deviceKeyset,
-                  [set],
-                  JSON.stringify(message),
-                  self,
-                  currentPasskeyInfo!.displayName,
-                  currentPasskeyInfo?.pfpUrl
-                ),
-              ];
-            } else {
-              sessions = [
-                ...sessions,
-                ...secureChannel.DoubleRatchetInboxEncrypt(
-                  keyset.deviceKeyset,
-                  [set],
-                  JSON.stringify(message),
-                  self,
-                  currentPasskeyInfo!.displayName,
-                  currentPasskeyInfo?.pfpUrl
-                ),
-              ];
-            }
-          } else {
-            sessions = [
-              ...sessions,
-              ...(await secureChannel.NewDoubleRatchetSenderSession(
-                keyset.deviceKeyset,
-                self.user_address,
-                self.device_registrations
-                  .concat(counterparty.device_registrations)
-                  .find((d) => d.inbox_registration.inbox_address === inbox)!,
-                JSON.stringify(message),
-                currentPasskeyInfo!.displayName,
-                currentPasskeyInfo?.pfpUrl
-              )),
-            ];
-          }
-        }
-
-        for (const session of sessions) {
-          const newEncryptionState: EncryptionState = {
-            state: JSON.stringify({
-              ratchet_state: session.ratchet_state,
-              receiving_inbox: session.receiving_inbox,
-              tag: session.tag,
-              sending_inbox: session.sending_inbox,
-            } as secureChannel.DoubleRatchetStateAndInboxKeys),
-            timestamp: Date.now(),
-            inboxId: session.receiving_inbox.inbox_address,
-            conversationId: address! + '/' + address!,
-            sentAccept: session.sent_accept,
-          };
-          await messageDB.saveEncryptionState(newEncryptionState, true);
-          outbounds.push(
-            JSON.stringify({
-              type: 'listen',
-              inbox_addresses: [session.receiving_inbox.inbox_address],
-            })
-          );
-          outbounds.push(
-            JSON.stringify({ type: 'direct', ...session.sealed_message })
-          );
-        }
-
-        // do not save delete-conversation message
-        if (message.content.type === 'delete-conversation') {
-          return outbounds;
-        }
-
-        await saveMessage(message, messageDB, address!, address!, 'direct', {
-          user_icon:
-            conversation?.conversation?.icon ?? DefaultImages.UNKNOWN_USER,
-          display_name:
-            conversation?.conversation?.displayName ?? t`Unknown User`,
-        });
-        await addMessage(queryClient, address, address, message);
-        addOrUpdateConversation(
-          queryClient,
-          address,
-          Date.now(),
-          message.createdDate,
-          {
-            user_icon:
-              conversation?.conversation?.icon ?? DefaultImages.UNKNOWN_USER,
-            display_name:
-              conversation?.conversation?.displayName ?? 'Unknown User',
-          }
-        );
-
-        return outbounds;
-      });
-    },
-    []
-  );
+  // submitMessage will be defined after messageService instantiation
 
   const int64ToBytes = (num: number) => {
     const arr = new Uint8Array(8);
@@ -5718,6 +3724,132 @@ const MessageDBProvider: FC<MessageDBContextProps> = ({ children }) => {
       }, 10000);
     }
   }, [keyset, selfAddress]);
+
+  // Create MessageService instance (after all dependencies are declared)
+  const messageService = useMemo(() => {
+    return new MessageService({
+      messageDB,
+      enqueueOutbound,
+      addOrUpdateConversation,
+      apiClient,
+      deleteEncryptionStates,
+      deleteInboxMessages,
+      navigate,
+      spaceInfo,
+      syncInfo,
+      synchronizeAll,
+      informSyncData,
+      initiateSync,
+      directSync,
+      saveConfig,
+      int64ToBytes,
+      canonicalize,
+    });
+  }, [messageDB, enqueueOutbound, apiClient, deleteEncryptionStates, deleteInboxMessages, navigate, spaceInfo, syncInfo, synchronizeAll, informSyncData, initiateSync, directSync, saveConfig, canonicalize]);
+
+  // START_HANDLE_NEW_MESSAGE_FUNCTION
+  // USING MessageService: handleNewMessage now delegates to the extracted service
+  const handleNewMessage = useCallback(
+    async (
+      self_address: string,
+      keyset: {
+        userKeyset: secureChannel.UserKeyset;
+        deviceKeyset: secureChannel.DeviceKeyset;
+      },
+      message: EncryptedMessage
+    ) => {
+      // Delegate to MessageService - maintaining exact same API
+      return messageService.handleNewMessage(
+        self_address,
+        keyset,
+        message,
+        queryClient
+      );
+    },
+    [messageService, queryClient]
+  );
+  // END_HANDLE_NEW_MESSAGE_FUNCTION
+
+  // START_SUBMIT_MESSAGE_FUNCTION
+  // USING MessageService: submitMessage now delegates to the extracted service
+  const submitMessage = React.useCallback(
+    async (
+      address: string,
+      pendingMessage: string | object,
+      self: secureChannel.UserRegistration,
+      counterparty: secureChannel.UserRegistration,
+      queryClient: QueryClient,
+      currentPasskeyInfo: {
+        credentialId: string;
+        address: string;
+        publicKey: string;
+        displayName?: string;
+        pfpUrl?: string;
+        completedOnboarding: boolean;
+      },
+      keyset: {
+        deviceKeyset: secureChannel.DeviceKeyset;
+        userKeyset: secureChannel.UserKeyset;
+      },
+      inReplyTo?: string,
+      skipSigning?: boolean
+    ) => {
+      // Delegate to MessageService - maintaining exact same API
+      return messageService.submitMessage(
+        address,
+        pendingMessage,
+        self,
+        counterparty,
+        queryClient,
+        currentPasskeyInfo,
+        keyset,
+        inReplyTo,
+        skipSigning
+      );
+    },
+    [messageService]
+  );
+  // END_SUBMIT_MESSAGE_FUNCTION
+
+  // START_SAVE_MESSAGE_FUNCTION
+  // USING MessageService: saveMessage now delegates to the extracted service
+  const saveMessage = async (
+    decryptedContent: Message,
+    messageDB: MessageDB,
+    spaceId: string,
+    channelId: string,
+    conversationType: string,
+    updatedUserProfile: { user_icon?: string; display_name?: string }
+  ) => {
+    // Delegate to MessageService - maintaining exact same API
+    return messageService.saveMessage(
+      decryptedContent,
+      messageDB,
+      spaceId,
+      channelId,
+      conversationType,
+      updatedUserProfile
+    );
+  };
+  // END_SAVE_MESSAGE_FUNCTION
+
+  // START_ADD_MESSAGE_FUNCTION
+  // USING MessageService: addMessage now delegates to the extracted service
+  const addMessage = async (
+    queryClient: QueryClient,
+    spaceId: string,
+    channelId: string,
+    decryptedContent: Message
+  ) => {
+    // Delegate to MessageService - maintaining exact same API
+    return messageService.addMessage(
+      queryClient,
+      spaceId,
+      channelId,
+      decryptedContent
+    );
+  };
+  // END_ADD_MESSAGE_FUNCTION
 
   return (
     <MessageDBContext.Provider
